producer/requirements.txt
kafka-python==2.0.2
pandas==1.5.2
selenium==4.7.2